+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 v2.0.2 ????
   * Add --user-agent option to set web User-Agent string
+  * Avoid misleading "(0 bytes)" output when no Content-Length given (https://github.com/resync/resync/issues/52)
 
 v2.0.1 2021-03-23
   * Route all URI and file requests through `resync/url_or_file_open.py` so that settings such as authentication headers can be consistently applied